backend/requirements.txt
fastapi
uvicorn
whisper
sonner
python-dotenv
psycopg2-binary
firebase-admin
requests